chore: rename PromptedInput to prompted-input
