+++ src/components/agent/agent-list/agent-item/agent-actions/add-resource-popover/prompted-input/index.test.tsx
